+++ subgraph/fluence-graph.sh
@@ -1,6 +1,5 @@
 #! /usr/bin/env bash
 
-# Note, only for local it uses prefix as full name to deploy subgraph.
 SUBGRAPH_NAME_PREFIX="fluence-deal-contracts"
 # Directory with saved deployments. It will append rows with other deployments.
 SUBGRAPH_DEPLOYMENTS_DIR="deployments"
@@ -9,14 +8,19 @@ help() {
 script_name="$(basename $0)"
 cat<<HELP
 Usage: [BASIC_AUTH_SUBGRAPH] ${script_name} network action
-Deploy subgraph to Fluence.
+Deploy subgraph to Fluence stands, or local. Store deployment artifacts (local is ignored).
 
   network       Fluence network to run against - local, stage, dar or kras
-  action        action to run - create, deploy or remove
+  action        action to run - create, deploy or delete
 
 Examples:
   BASIC_AUTH_SUBGRAPH=user:pass ${script_name} stage deploy     Deploy subgraph to stage
-  ${script_name} local deploy                                   Deploy subgraph to local
+  ${script_name} local deploy                                   Deploy subgraph to local (BASIC_AUTH_SUBGRAPH not used)
+
+Envs:
+  - GRAPHNODE_ADMIN_URL_LOCAL: url for local network admin graph panel.
+  - GRAPHNODE_QUERY_URL_LOCAL: url for local network graph query
+  - SUBGRAPH_NAME_TO_DELETE: e.g. for local fluence-deal-contracts
 HELP
 }
 
@@ -26,8 +30,16 @@ HELP
 # Parse script arguments
 while (($#)); do
   case "$1" in
-    deploy|create|remove)
+    deploy|create)
       action="$1"
+      shift
+      ;;
+    delete)
+      action="$1"
+      if [[ -z $SUBGRAPH_NAME_TO_DELETE ]]; then
+          echo "For delete action SUBGRAPH_NAME_TO_DELETE is required."
+          exit 1
+      fi
       shift
       ;;
     local|stage|kras|dar)
@@ -48,23 +60,10 @@ done
 # Prepare variables depending on network
 case "$network" in
   local)
-    GRAPHNODE_URL="${GRAPHNODE_URL:-http://localhost:8020}"
+    GRAPHNODE_URL="${GRAPHNODE_ADMIN_URL_LOCAL:-http://localhost:8020}"
     IPFS_URL="${IPFS_URL:-http://localhost:5001}"
     ;;
-  stage)
-    if [[ -z $BASIC_AUTH_SUBGRAPH ]]; then
-      echo "Please provide credentials with 'BASIC_AUTH_SUBGRAPH' variable."
-      exit 1
-    else
-      basic_auth="${BASIC_AUTH_SUBGRAPH}@"
-    fi
-    GRAPHNODE_URL="https://${basic_auth}graph-node-admin-${network}.fluence.dev"
-    # To save logs of created subgraphs on Fluence stands.
-    GRAPH_NODE_QUERY_URL="https://graph-node-${network}.fluence.dev/subgraphs/name/"
-    # TODO: IPFS why does not work with auth (before deploy request Tolay to open ports)
-    IPFS_URL="https://graph-node-ipfs-${network}.fluence.dev"
-    ;;
-  kras|dar)
+  stage|kras|dar)
     if [[ -z $BASIC_AUTH_SUBGRAPH ]]; then
       echo "Please provide credentials with 'BASIC_AUTH_SUBGRAPH' variable."
       exit 1
@@ -72,13 +71,12 @@ case "$network" in
       basic_auth="${BASIC_AUTH_SUBGRAPH}@"
     fi
     GRAPHNODE_URL="https://${basic_auth}graph-node-admin.${network}.fluence.dev"
-    # To save logs of created subgraphs on Fluence stands.
-    GRAPH_NODE_QUERY_URL="https://graph-node.${network}.fluence.dev/subgraphs/name/"
-    # TODO: IPFS why does not work with auth (before deploy request Tolay to open ports)
     IPFS_URL="https://graph-node-ipfs.${network}.fluence.dev"
 esac
 
 # Prepare subgraph version label.
+# Do we need this [currently no]?
+#  ref to https://github.com/graphprotocol/graph-node/pull/696
 SUBGRAPH_VERSION_LABEL="${SUBGRAPH_VERSION_LABEL:-0.0.0}"
 
 # Wait for graph-node to be up
@@ -87,7 +85,7 @@ echo "Trying to connect to Graph Node at ${GRAPHNODE_URL}..."
 while [[ $retries > 0 ]]; do
   status_code=$(curl -o /dev/null -s -w "%{http_code}\n" -X POST -H "Content-Type: application/json" -d '{"foo":"bar"}' ${GRAPHNODE_URL})
   if [[ $status_code == "200" ]]; then
-    echo "Succesfully connected to Graph Node."
+    echo "Successfully connected to Graph Node. Sleep before start 10 seconds."
     sleep 10
     break
   else
@@ -120,30 +118,46 @@ case "$network" in
     ;;
 esac
 
-echo "Create unique subgraph name for the current state of contract deployments via git commit hash (ignore this suffix for local)..."
 case "$network" in
-  local|stage|kras|dar)
+  local)
     SUBGRAPH_NAME="${SUBGRAPH_NAME_PREFIX}"
     ;;
-#  TODO: resolve: do we need versioning of subgraph or not finally.
-#  stage|kras|dar)
-#    SUBGRAPH_NAME="${SUBGRAPH_NAME_PREFIX}-$(git rev-parse --short HEAD)"
-#    ;;
+  stage|kras|dar)
+    echo "Prepare deployment log (append url to deployments with commit hash specified)..."
+    SUBGRAPH_NAME="${SUBGRAPH_NAME_PREFIX}-$(git rev-parse --short HEAD)"
+    ;;
 esac
 
+DEPLOYMENT_LOG_PATH="${SUBGRAPH_DEPLOYMENTS_DIR}/${network}.txt"
+# Need to exit if not deployed successfully (to stop update artifacts).
+set -e
 case "$action" in
   deploy)
-    echo "Deploying subgraph on ${network} stand with subgraph name: $SUBGRAPH_NAME and version label $SUBGRAPH_VERSION_LABEL..."
+    echo "Deploying subgraph on ${network} stand with subgraph name: ${SUBGRAPH_NAME} and version label ${SUBGRAPH_VERSION_LABEL}..."
     auth_header=$(echo -n ${BASIC_AUTH_SUBGRAPH} | base64)
-    graph deploy --node ${GRAPHNODE_URL} --headers "{\"Authorization\": \"Basic ${auth_header}\"}" --ipfs ${IPFS_URL} --network ${SUBGRAPH_NETWORK} --network-file configs/${network}-networks-config.json --version-label ${SUBGRAPH_VERSION_LABEL} ${SUBGRAPH_NAME} \
+    graph deploy --node ${GRAPHNODE_URL} --headers "{\"Authorization\": \"Basic ${auth_header}\"}" --ipfs ${IPFS_URL} --network ${SUBGRAPH_NETWORK} --network-file configs/${network}-networks-config.json --version-label ${SUBGRAPH_VERSION_LABEL} ${SUBGRAPH_NAME}
+    # Finally store logs if needed.
+    if [[ $network == "local" ]]; then
+      echo "Update of artifacts of local network is ignored."
+    else
+      echo "Append deployment URL: ${SUBGRAPH_NAME} into ${DEPLOYMENT_LOG_PATH}..."
+      echo "${SUBGRAPH_NAME}" >> ${DEPLOYMENT_LOG_PATH}
+    fi
     ;;
   create)
-    echo "Creating subgraph on Fluence with name: $SUBGRAPH_NAME..."
+    echo "Creating subgraph on ${network} with name: ${SUBGRAPH_NAME}..."
     graph create --node ${GRAPHNODE_URL} ${SUBGRAPH_NAME}
-    echo "${GRAPH_NODE_QUERY_URL}${SUBGRAPH_NAME}" >> ${SUBGRAPH_DEPLOYMENTS_DIR}/${network}.txt
     ;;
-  remove)
-    echo "Removing subgraph on Fluence with name: $SUBGRAPH_NAME..."
-    graph remove --node ${GRAPHNODE_URL} ${SUBGRAPH_NAME}
+  delete)
+    echo "Deleting subgraph on ${network} network with name: ${SUBGRAPH_NAME_TO_DELETE} from ${GRAPHNODE_URL}..."
+    graph remove --node ${GRAPHNODE_URL} ${SUBGRAPH_NAME_TO_DELETE}
+    if [[ $network == "local" ]]; then
+      echo "Update of artifacts of local network is ignored."
+    else
+      echo "Update artifacts in: ${DEPLOYMENT_LOG_PATH}"
+      sed -e "s/${SUBGRAPH_NAME_TO_DELETE}//g" -i '' ${DEPLOYMENT_LOG_PATH}
+      # Delete empty lines as well.
+      sed -e '/^$/d' -i '' ${DEPLOYMENT_LOG_PATH}
+    fi
     ;;
 esac